drivers: modem: cellular: Move driver macros into internal header
To allow cellular modems to be declared outside of modem_cellular.c
file, move all structure definitions and macros into internal header
zephyr/drivers/cellular_internal.h

This allows out-of-tree modem to be defined with
MODEM_CELLULAR_DEFINE_INSTANCE() macro.

Remove Devicetree dependency from CONFIG_MODEM_CELLULAR
