+++ include/zephyr/modem/ppp.h
@@ -8,6 +8,7 @@
 #include <zephyr/types.h>
 #include <zephyr/net/net_if.h>
 #include <zephyr/net/net_pkt.h>
+#include <zephyr/net/ppp.h>
 #include <zephyr/sys/ring_buffer.h>
 #include <zephyr/sys/atomic.h>
